+++ tencentcloud/ckafka/v20190819/models.go
@@ -1320,6 +1320,25 @@ func (r *DescribeUserResponse) FromJsonString(s string) error {
     return json.Unmarshal([]byte(s), &r)
 }
 
+type DynamicRetentionTime struct {
+
+	// 动态消息保留时间配置开关（0: 关闭，1: 开启）
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	Enable *int64 `json:"Enable,omitempty" name:"Enable"`
+
+	// 磁盘配额百分比触发条件，即消息达到此值触发消息保留时间变更事件
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	DiskQuotaPercentage *int64 `json:"DiskQuotaPercentage,omitempty" name:"DiskQuotaPercentage"`
+
+	// 每次向前调整消息保留时间百分比
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	StepForwardPercentage *int64 `json:"StepForwardPercentage,omitempty" name:"StepForwardPercentage"`
+
+	// 保底时长，单位分钟
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	BottomRetention *int64 `json:"BottomRetention,omitempty" name:"BottomRetention"`
+}
+
 type Filter struct {
 
 	// 需要过滤的字段。
@@ -1537,7 +1556,7 @@ type InstanceAttributesResponse struct {
 	// 注意：此字段可能返回 null，表示取不到有效值。
 	MaxGroupNum *int64 `json:"MaxGroupNum,omitempty" name:"MaxGroupNum"`
 
-	// 售卖类型
+	// 售卖类型,0:标准版,1:专业版
 	// 注意：此字段可能返回 null，表示取不到有效值。
 	Cvm *int64 `json:"Cvm,omitempty" name:"Cvm"`
 
@@ -1548,6 +1567,10 @@ type InstanceAttributesResponse struct {
 	// 表示该实例支持的特性。FEATURE_SUBNET_ACL:表示acl策略支持设置子网。
 	// 注意：此字段可能返回 null，表示取不到有效值。
 	Features []*string `json:"Features,omitempty" name:"Features" list`
+
+	// 动态消息保留策略
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	RetentionTimeConfig *DynamicRetentionTime `json:"RetentionTimeConfig,omitempty" name:"RetentionTimeConfig"`
 }
 
 type InstanceConfigDO struct {
@@ -1632,6 +1655,26 @@ type InstanceDetail struct {
 	// ckafka售卖类型
 	// 注意：此字段可能返回 null，表示取不到有效值。
 	Cvm *int64 `json:"Cvm,omitempty" name:"Cvm"`
+
+	// ckafka实例类型
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	InstanceType *string `json:"InstanceType,omitempty" name:"InstanceType"`
+
+	// 磁盘类型
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	DiskType *string `json:"DiskType,omitempty" name:"DiskType"`
+
+	// 当前规格最大Topic数
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	MaxTopicNumber *int64 `json:"MaxTopicNumber,omitempty" name:"MaxTopicNumber"`
+
+	// 当前规格最大Partition数
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	MaxPartitionNumber *int64 `json:"MaxPartitionNumber,omitempty" name:"MaxPartitionNumber"`
+
+	// 计划升级配置时间
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	RebalanceTime *string `json:"RebalanceTime,omitempty" name:"RebalanceTime"`
 }
 
 type InstanceDetailResponse struct {
@@ -1751,6 +1794,12 @@ type ModifyInstanceAttributesRequest struct {
 
 	// 实例配置
 	Config *ModifyInstanceAttributesConfig `json:"Config,omitempty" name:"Config"`
+
+	// 动态消息保留策略配置
+	DynamicRetentionConfig *DynamicRetentionTime `json:"DynamicRetentionConfig,omitempty" name:"DynamicRetentionConfig"`
+
+	// 修改升配置rebalance时间
+	RebalanceTime *int64 `json:"RebalanceTime,omitempty" name:"RebalanceTime"`
 }
 
 func (r *ModifyInstanceAttributesRequest) ToJsonString() string {
@@ -2059,6 +2108,10 @@ type TopicDetail struct {
 	// 高级配置
 	// 注意：此字段可能返回 null，表示取不到有效值。
 	Config *Config `json:"Config,omitempty" name:"Config"`
+
+	// 消息保留时间配置(用于动态配置变更记录)
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	RetentionTimeConfig *TopicRetentionTimeConfigRsp `json:"RetentionTimeConfig,omitempty" name:"RetentionTimeConfig"`
 }
 
 type TopicDetailResponse struct {
@@ -2095,6 +2148,21 @@ type TopicResult struct {
 	// 符合条件的 topic 数量
 	// 注意：此字段可能返回 null，表示取不到有效值。
 	TotalCount *int64 `json:"TotalCount,omitempty" name:"TotalCount"`
+}
+
+type TopicRetentionTimeConfigRsp struct {
+
+	// 期望值，即用户配置的Topic消息保留时间(单位分钟)
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	Expect *int64 `json:"Expect,omitempty" name:"Expect"`
+
+	// 当前值，即当前生效值(可能存在动态调整，单位分钟)
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	Current *int64 `json:"Current,omitempty" name:"Current"`
+
+	// 最近变更时间
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	ModTimeStamp *int64 `json:"ModTimeStamp,omitempty" name:"ModTimeStamp"`
 }
 
 type User struct {